add cdn profile facts module - doc TBD
